Add modules. Install webpack
