+++ src/utils.js
@@ -1,0 +1,3 @@
+const generateRandomNumber = (min, max) => Math.floor(Math.random() * (max - min) + min);
+
+export {generateRandomNumber};
